Moved directories into separate file
